src/requirements.txt
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
boto==2.48.0
boto3==1.7.1
botocore==1.10.1
cairocffi==0.8.0
CairoSVG==2.1.3
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
decorator==4.0.10
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11
django-appconf==1.0.2
django-crispy-forms==1.7.1
django-filter==1.1.0
django-imagekit==4.0.2
django-material==1.2.2
django-paypal==0.5.0
django-phonenumber-field==2.0.0
django-rest-framework==0.1.0
django-storages==1.6.6
djangorestframework==3.7.7
docutils==0.14
enum34==1.1.6
gunicorn==19.7.1
html5lib==1.0.1
idna==2.6
instamojo-wrapper==1.1
ipython-genutils==0.1.0
jmespath==0.9.3
pathlib2==2.1.0
pdfrw==0.4
pexpect==4.2.1
phonenumberslite==8.9.4
pickleshare==0.7.4
pilkit==2.0
Pillow==5.0.0
prompt-toolkit==1.0.9
psycopg2==2.7.4
ptyprocess==0.5.1
pycparser==2.18
Pyphen==0.9.4
python-dateutil==2.6.1
pytz==2018.3
razorpay==1.1.1
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
stripe==1.79.1
tinycss2==0.6.1
urllib3==1.22
wcwidth==0.1.7
WeasyPrint==0.42.3
webencodings==0.5.1
